plugins: logterse.pl
Charles Butcher 2007-02-27 08:53:55 UTC
https://qpsmtpd.perl.narkive.com/Fi7GnV78/annonuncing-logterse-a-new-plugin

Annonuncing logterse - a new plugin

Here is a new plugin that provides concise, informative log entries for
each message qpsmtpd processes. It is accompanied by a log analysis
tool that takes advantage of multilog's features, although you do not
_have_ to use it with multilog.

The plugin and its analysis tool are available for the time being at
http://ncc.com.au/qpsmtpd/logterse

How to use it:

- invoke logterse from the config/plugins file the same as for any
plugin.